% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{catalog_summary}
\alias{catalog_summary}
\title{Print catalog summary}
\usage{
catalog_summary()
}
\value{
NULL (prints summary to console)
}
\description{
Print catalog summary
}
\examples{
catalog_summary()
}
