% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_data_types}
\alias{list_data_types}
\title{List available data types}
\usage{
list_data_types()
}
\value{
Character vector of data type names
}
\description{
List available data types
}
\examples{
list_data_types()
}
