% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_dimensions}
\alias{list_dimensions}
\title{List available dimensions}
\usage{
list_dimensions()
}
\value{
Character vector of dimension names
}
\description{
List available dimensions
}
\examples{
list_dimensions()
}
