% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_domains}
\alias{list_domains}
\title{List available WRI domains}
\usage{
list_domains()
}
\value{
Character vector of domain names
}
\description{
List available WRI domains
}
\examples{
list_domains()
}
