% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_stac.R
\name{query_stac_flexible}
\alias{query_stac_flexible}
\title{Query STAC items flexibly by any property}
\usage{
query_stac_flexible(items, ...)
}
\arguments{
\item{items}{List of STAC items}

\item{...}{Named property-value pairs to filter by}
}
\value{
Filtered list of items
}
\description{
Query STAC items flexibly by any property
}
\examples{
items <- list(
  list(properties = list(wri_domain = "water", data_type = "status")),
  list(properties = list(wri_domain = "land", data_type = "status"))
)
query_stac_flexible(items, wri_domain = "water", data_type = "status")
}
