% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wri_overview.R
\name{wri_overview}
\alias{wri_overview}
\title{Read the WRI STAC catalog}
\usage{
wri_overview(fresh = TRUE)
}
\arguments{
\item{fresh}{Logical. If TRUE (default), rebuilds the catalog from disk.
If FALSE, not implemented yet (reserved for caching).}
}
\value{
An object of class `wri_overview`.
}
\description{
Reads the local STAC catalog and returns a `wri_overview` object.
}
