% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer.R
\name{get_layer}
\alias{get_layer}
\title{Get a WRI raster layer}
\usage{
get_layer(id, bbox = NULL, crs = "EPSG:4326")
}
\arguments{
\item{id}{Character. Layer id to retrieve.}

\item{bbox}{Numeric vector \code{c(xmin, ymin, xmax, ymax)} in the supplied
\code{crs}. If \code{NULL}, returns the full layer.}

\item{crs}{Character. Coordinate reference system of \code{bbox}. Defaults to
\code{"EPSG:4326"}.}
}
\value{
A \code{terra::SpatRaster}.
}
\description{
Get a WRI raster layer
}
\examples{
\dontrun{
bbox <- c(-122, 37, -121, 38)
rast <- get_layer("WRI_score", bbox = bbox)
}
}
